requirements.txt
opencv-python>=4.8
mediapipe==0.10.11
numpy>=1.24,<2.0
matplotlib>=3.7
scipy>=1.10
pyyaml>=6.0.1